refactor: changement noms des stores
