+++ src/components/level-component.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react'
 import ButtonWrapper from '@/components/button/button-cercle'
 import Avatar from '@/components/card/card-avatar'
 import CardRewardLevel from '@/components/card/card-reward-level'
-import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
-import useLevelStore from '@/store/level-progress'
+import useTotalQuizStore from '@/store/tracking/total-answer'
+import useLevelStore from '@/store/tracking/level-progress'
 import CardCode from '@/components/card/card-code-user'
 
 export interface LevelComponentProps {
