+++ src/pages/completed.tsx
@@ -3,7 +3,7 @@ import ContentCorrectAnswer from '@/components/card/card-correct-answer'
 import CardReward from '@/components/card/card-reward'
 import CustomButton from '@/components/button/button-long'
 import { useNavigate } from 'react-router-dom'
-import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
+import useTotalQuizStore from '@/store/tracking/total-answer'
 import { ROUTES } from '@/const'
 import useSound from 'use-sound'
 import { useEffect } from 'react'
