+++ src/pages/end.tsx
@@ -1,10 +1,10 @@
 import CustomButton from '@/components/button/button-long'
 import Picture from '@/components/picture'
-import useLevelStore from '@/store/level-progress'
-import useProgressStore from '@/store/tracking/tracker-progress'
-import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
+import useLevelStore from '@/store/tracking/level-progress'
+import useProgressStore from '@/store/tracking/exercise-progress'
+import useTotalQuizStore from '@/store/tracking/total-answer'
 import { useNavigate } from 'react-router-dom'
-import useAuthStore from '@/store/tracking/tracker-auth'
+import useAuthStore from '@/store/tracking/auth'
 import { API_URL } from '@/config/api'
 import { ROUTES } from '@/const'
