+++ src/pages/result.tsx
@@ -2,10 +2,10 @@ import Title from '@/components/title'
 import ContentCorrectAnswer from '@/components/card/card-correct-answer'
 import CardReward from '@/components/card/card-reward'
 import CustomButton from '@/components/button/button-long'
-import useQuizStore from '@/store/tracking/tracker-answer'
-import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
-import useProgressStore from '@/store/tracking/tracker-progress'
-import useLevelStore from '@/store/level-progress'
+import useQuizStore from '@/store/tracking/answer'
+import useTotalQuizStore from '@/store/tracking/total-answer'
+import useProgressStore from '@/store/tracking/exercise-progress'
+import useLevelStore from '@/store/tracking/level-progress'
 import { useNavigate } from 'react-router-dom'
 import { useEffect, useRef } from 'react'
 import { ROUTES } from '@/const'
